requirements.txt
youtube-transcript-api==0.6.2
yt-dlp==2024.4.9
tqdm==4.66.4
openai==1.30.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==2.0.1